+++ assets/css/index.ts
@@ -2,3 +2,4 @@ import './primercss.scss';
 import './actionlist/index.scss';
 import './prompt.css';
 import './primercss-overrides.css';
+import './toggle-switch.css';