+++ test/components/select_menu_test.exs
@@ -315,7 +315,7 @@ defmodule PrimerLive.TestComponents.SelectMenuTest do
              <div class="SelectMenu">
              <div aria-role="menu" class="SelectMenu-modal" data-content="">
              <div class="SelectMenu-loading"><svg class="octicon anim-pulse" xmlns="http://www.w3.org/2000/svg" width="48"
-             height="48" viewBox="0 0 48 48">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg></div>
+             height="48" viewBox="0 0 48 48">STRIPPED_SVG_PATHS</svg></div>
              <div class="SelectMenu-list"></div>
              </div>
              </div>
@@ -565,7 +565,7 @@ defmodule PrimerLive.TestComponents.SelectMenuTest do
              role="tab">Selected tab</button><button class="SelectMenu-tab tab-x" role="tab">Other tab</button></div>
              <div class="SelectMenu-loading loading-x my-loading"><svg class="octicon anim-pulse"
              xmlns="http://www.w3.org/2000/svg" width="48" height="48"
-             viewBox="0 0 48 48">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg></div>
+             viewBox="0 0 48 48">STRIPPED_SVG_PATHS</svg></div>
              <div class="SelectMenu-list menu_list-x">
              <div class="SelectMenu-blankslate blankslate-x my-blankslate">Blankslate content</div><button
              class="SelectMenu-item item-x my-item" role="menuitem">Item 1</button><button class="SelectMenu-item item-x"