+++ test/components/spinner_test.exs
@@ -14,7 +14,8 @@ defmodule PrimerLive.Components.SpinnerTest do
            """)
            |> format_html() ==
              """
-             <svg class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg>
+             <svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none"
+             class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg>
              """
              |> format_html()
   end
@@ -28,8 +29,10 @@ defmodule PrimerLive.Components.SpinnerTest do
            """)
            |> format_html() ==
              """
-             <svg class="Toast--spinner" height="40" viewBox="0 0 32 32" width="40">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg><svg
-             class="Toast--spinner" height="40" viewBox="0 0 32 32" width="40">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg>
+             <svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none"
+             class="anim-rotate" height="40" width="40">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16"
+             xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none" class="anim-rotate" height="40"
+             width="40">STRIPPED_SVG_PATHS</svg>
              """
              |> format_html()
   end
@@ -44,9 +47,11 @@ defmodule PrimerLive.Components.SpinnerTest do
            """)
            |> format_html() ==
              """
-             <svg class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg><svg
-             class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg><svg
-             class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg>
+             <svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none"
+             class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16"
+             xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none" class="anim-rotate" height="32"
+             width="32">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true"
+             focusable="false" fill="none" class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg>
              """
              |> format_html()
   end
@@ -61,9 +66,11 @@ defmodule PrimerLive.Components.SpinnerTest do
            """)
            |> format_html() ==
              """
-             <svg class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg><svg
-             class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg><svg
-             class="Toast--spinner" height="18" viewBox="0 0 32 32" width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg>
+             <svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none"
+             class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16"
+             xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none" class="anim-rotate" height="32"
+             width="32">STRIPPED_SVG_PATHS</svg><svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true"
+             focusable="false" fill="none" class="anim-rotate" height="32" width="32">STRIPPED_SVG_PATHS</svg>
              """
              |> format_html()
   end
@@ -76,8 +83,8 @@ defmodule PrimerLive.Components.SpinnerTest do
            """)
            |> format_html() ==
              """
-             <svg class="Toast--spinner my-spinner" dir="rtl" height="18" viewBox="0 0 32 32"
-             width="18">STRIPPED_SVG_PATHSSTRIPPED_SVG_PATHS</svg>
+             <svg viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" fill="none"
+             class="anim-rotate my-spinner" dir="rtl" height="32" width="32" dir="rtl">STRIPPED_SVG_PATHS</svg>
              """
              |> format_html()
   end