+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Component `avatar_pair`: renamed `parent_child_avatar` to `avatar_pair`.
 - Components `action_menu` and `select_menu`: renamed `is_right_aligned` to `is_aligned_end` (added RTL support).
 
+### New component
+
+- `toggle_switch`: Toggle switch is used to immediately toggle a setting on or off.
+
 ### Updated components
 
 - `avatar`: Added attr `is_round`.